pyproject.toml
[tool.poetry]
name = "ermlab"
version = "0.1.0"
description = "Api pozycz i zapomnij"
authors = ["Kamil Cholipski <kamil.cholipski@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.6"
Django = "^3.1.7"
djangorestframework = "^3.12.2"
psycopg2 = "^2.8.6"
djangorestframework-simplejwt = "^4.6.0"
django-cors-headers = "^3.7.0"
asgiref = "^3.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
